requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
aiostream
altair
annotated-types
anyio
appnope
asgiref
asttokens
async-timeout
attrs
backcall
backoff
bcrypt
beautifulsoup4
bleach
blinker
blis
bm25s
bs4
cachetools
catalogue
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
cloudpathlib
coloredlogs
comm
confection
contourpy
cryptography
cycler
cymem
dataclasses-json
debugpy
decorator
defusedxml
Deprecated
dirtyjson
distro
docopt
docx2txt
durationpy
easyocr
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup
executing
fastapi
fastjsonschema
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
gitdb
GitPython
google-auth
googleapis-common-protos
greenlet
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httptools
httpx
huggingface-hub
humanfriendly
hyperframe
idna
imageio
importlib-metadata
importlib_resources
ipykernel
ipython
jedi
Jinja2
jiter
joblib
jpype1
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
kiwisolver
kubernetes
langcodes
language_data
lazy_loader
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index==0.12.24
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.24.post1
llama-index-embeddings-huggingface==0.5.3
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-legacy==0.9.48.post4
llama-index-llms-huggingface==0.4.2
llama-index-llms-openai==0.3.25
llama-index-llms-openai-like==0.3.4
llama-index-llms-openrouter==0.3.1
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.2
llama-parse==0.6.4.post1
llamaindex-py-client==0.1.19
marisa-trie
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mistune
mmh3
monotonic
mpmath
multidict
murmurhash
mypy-extensions
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
ninja
nltk
numpy
oauthlib
onnxruntime
openai
opencv-python-headless
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
overrides
packaging
pandas
pandocfilters
parso
pdf2image
pdfminer.six
pdfplumber
pexpect
pickleshare
pillow
pipreqs
platformdirs
plotly
portalocker
posthog
preshed
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pulsar-client
pure_eval
pyarrow
pyasn1
pyasn1_modules
pyclipper
pycparser
pydantic
pydantic_core
pydeck
Pygments
PyMuPDF
pyparsing
pypdf
pypdfium2
PyPika
pyspellchecker
PyStemmer
pytesseract
python-bidi
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
qdrant-client
referencing
regex
requests
requests-oauthlib
rich
rpds-py
rsa
safetensors
scikit-image
scikit-learn
scipy
seaborn
sentence-transformers
shapely
shellingham
six
smart-open
smmap
sniffio
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
stack-data
starlette
streamlit
striprtf
sympy
tabula-py
tabulate
tenacity
text-generation
thinc
threadpoolctl
tifffile
tiktoken
tinycss2
tokenizers
toml
torch
torchvision
tornado
tqdm
traitlets
transformers
typer
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
uvloop
validators
wasabi
watchfiles
wcwidth
weasel
webencodings
websocket-client
websockets
wrapt
yarg
yarl
zipp